Clean CLI API
