+++ .gitignore
@@ -6,3 +6,5 @@ output/
 .vscode/
 ignore/
 .nyc_output/
+build/
+prec-*.tgz
